+++ packages/component-vue3/src/index.ts
@@ -1,0 +1,5 @@
+import Phormal from "./phormal-vue3";
+
+export {
+  Phormal,
+}