+++ packages/theme-basic/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0-alpha.10](https://github.com/tomosterlund/phormal/compare/v0.1.0-alpha.9...v0.1.0-alpha.10) (2023-01-02)
+
+**Note:** Version bump only for package @phormal/theme-basic
+
+
+
+
+
+# [0.1.0-alpha.9](https://github.com/tomosterlund/phormal/compare/v0.1.0-alpha.8...v0.1.0-alpha.9) (2023-01-01)
+
+**Note:** Version bump only for package @phormal/theme-basic
+
+
+
+
+
+# [0.1.0-alpha.8](https://github.com/tomosterlund/phormal/compare/v0.1.0-alpha.7...v0.1.0-alpha.8) (2023-01-01)
+
+**Note:** Version bump only for package @phormal/theme-basic
+
+
+
+
+
+# [0.1.0-alpha.7](https://github.com/tomosterlund/super-form/compare/v0.1.0-alpha.6...v0.1.0-alpha.7) (2023-01-01)
+
+**Note:** Version bump only for package @phormal/theme-basic
+
+
+
+
+
 # [0.1.0-alpha.6](https://github.com/tomosterlund/super-form/compare/v0.1.0-alpha.5...v0.1.0-alpha.6) (2023-01-01)