+++ packages/use-required/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0-alpha.10](https://github.com/tomosterlund/phormal/compare/v0.1.0-alpha.9...v0.1.0-alpha.10) (2023-01-02)
+
+**Note:** Version bump only for package @phormal/use-required
+
+
+
+
+
+# [0.1.0-alpha.9](https://github.com/tomosterlund/phormal/compare/v0.1.0-alpha.8...v0.1.0-alpha.9) (2023-01-01)
+
+**Note:** Version bump only for package @phormal/use-required
+
+
+
+
+
+# [0.1.0-alpha.8](https://github.com/tomosterlund/phormal/compare/v0.1.0-alpha.7...v0.1.0-alpha.8) (2023-01-01)
+
+**Note:** Version bump only for package @phormal/use-required
+
+
+
+
+
 # [0.1.0-alpha.4](https://github.com/tomosterlund/super-form/compare/v0.1.0-alpha.3...v0.1.0-alpha.4) (2022-12-29)
 
 **Note:** Version bump only for package @phormal/use-required